image_info/pyproject.toml
[project]
name = "image-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "rich>=14.2.0",
]
